+++ lib/screen/task_screen.dart
@@ -134,7 +134,7 @@ class _TaskScreenState extends ConsumerState<TaskScreen> {
                     width: 0,
                   ),
             Visibility(
-              visible: true,
+              visible: false,
               child: IconButton(
                 onPressed: () async {
                   // progressDialogue();